LFU-748 power: regulator: Fix power on/off delay issue
We meet SD initialization issue with some UHS-I SD cards on
iMX8MM/iMX93/iMX91 EVK. When sending operation condition to card,
the OCR does not return correct status. We find it is the issue
in MMC power cycle after below commit applied:
4fcba5d (regulator: implement basic reference counter)

When SD startup, the sequence of MMC power cycle is:
mmc_power_init(get vmmc_supply dev) -> mmc_power_off -> udelay(2000)
-> mmc_power_on

Before the commit, as a fixed regulator, the GPIO is set as:
  GPIO inactive (in mmc_power_init) ->
  GPIO inactive and delay off-on-delay-us (in mmc_power_off) ->
  udelay(2000) ->
  GPIO active (in mmc_power_on)

After the commit:
  GPIO inactive (in mmc_power_init) ->
  enable_count is 0, regulator_set_enable returns -EALREADY,
  GPIO is inactive but NO delay (in mmc_power_off) ->
  udelay(2000) ->
  GPIO active (in mmc_power_on)

Because the lost of off-on-delay-us, the SD card is not completely
power off. To fix the issue, add the delay after the GPIO setting in
regulator_common_of_to_plat which is called in device probing.
So in mmc_power_init, after GPIO is set to default inactive, the
off-on-delay-us is applied.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
(cherry picked from commit bf506bcc7800f1b2cbed89aa64bc7363c96a8bbd)